+++ cmd/broker/suite_test.go
@@ -4,12 +4,11 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/kyma-project/kyma-environment-broker/internal/metricsv2"
-
 	"github.com/google/uuid"
 	"github.com/kyma-project/kyma-environment-broker/common/gardener"
 	pkg "github.com/kyma-project/kyma-environment-broker/common/runtime"
 	"github.com/kyma-project/kyma-environment-broker/internal/broker"
+	"github.com/kyma-project/kyma-environment-broker/internal/metrics"
 	"github.com/kyma-project/kyma-environment-broker/internal/process"
 
 	"github.com/kyma-project/kyma-environment-broker/internal/storage"
@@ -284,7 +283,7 @@ func fixConfig() *Config {
 		Deprovisioning:                            process.StagedManagerConfiguration{MaxStepProcessingTime: time.Minute},
 		Update:                                    process.StagedManagerConfiguration{MaxStepProcessingTime: time.Minute},
 		UpdateRuntimeResourceDelay:                time.Millisecond,
-		MetricsV2: metricsv2.Config{
+		Metrics: metrics.Config{
 			Enabled:                                         true,
 			OperationResultRetentionPeriod:                  time.Hour,
 			OperationResultPollingInterval:                  3 * time.Second,
