+++ internal/metrics/metrics.go
@@ -1,4 +1,4 @@
-package metricsv2
+package metrics
 
 import (
 	"context"
@@ -19,7 +19,7 @@ import (
 const (
 	prometheusNamespacev2 = "kcp"
 	prometheusSubsystemv2 = "keb_v2"
-	logPrefix             = "@metricsv2"
+	logPrefix             = "@metrics"
 )
 
 // Exposer gathers metrics and keeps these in memory and exposes to prometheus for fetching, it gathers them by:
@@ -49,8 +49,8 @@ type RegisterContainer struct {
 
 func Register(ctx context.Context, sub event.Subscriber, db storage.BrokerStorage, cfg Config, logger *slog.Logger) *RegisterContainer {
 	logger = logger.With("from:", logPrefix)
-	logger.Info("Registering metricsv2")
-	logger.Info(fmt.Sprintf("MetricsV2 Config: Enabled=%v, OperationResultRetentionPeriod=%v, OperationResultPollingInterval=%v, OperationStatsPollingInterval=%v, OperationResultFinishedOperationRetentionPeriod=%v, BindingsStatsPollingInterval=%v",
+	logger.Info("Registering metrics")
+	logger.Info(fmt.Sprintf("Metrics Config: Enabled=%v, OperationResultRetentionPeriod=%v, OperationResultPollingInterval=%v, OperationStatsPollingInterval=%v, OperationResultFinishedOperationRetentionPeriod=%v, BindingsStatsPollingInterval=%v",
 		cfg.Enabled, cfg.OperationResultRetentionPeriod, cfg.OperationResultPollingInterval, cfg.OperationStatsPollingInterval, cfg.OperationResultFinishedOperationRetentionPeriod, cfg.BindingsStatsPollingInterval))
 	opDurationCollector := NewOperationDurationCollector(logger)
 	prometheus.MustRegister(opDurationCollector)
