+++ internal/metrics/operations_stats.go
@@ -1,4 +1,4 @@
-package metricsv2
+package metrics
 
 import (
 	"context"
